refactor(cli): split command argument modules
