+++ crates/fabrik-cli/src/cli.rs
@@ -4,8 +4,19 @@ use std::path::PathBuf;
 use std::process::ExitCode;
 
 use clap::{Parser, Subcommand, ValueEnum};
-use fabrik_cas::Digest;
 use fabrik_core::WorkspacePath;
+
+mod auth;
+mod cache;
+mod deps;
+mod runtime;
+mod toolchain;
+
+pub use auth::AuthCmd;
+pub use cache::{CacheActionCmd, CacheBlobCmd, CacheCmd};
+pub use deps::DepsCmd;
+pub use runtime::RuntimeCmd;
+pub use toolchain::ToolchainCmd;
 
 /// Workspace-relative directory holding Fabrik's CAS, action results,
 /// and build outputs. Hidden so VCS and editors ignore it by default.
@@ -287,140 +298,6 @@ pub enum Cmd {
     },
 }
 
-#[derive(Subcommand)]
-pub enum CacheCmd {
-    /// Print blob and action counts plus on-disk size.
-    Stats,
-
-    /// Read and write content-addressed blobs.
-    #[command(arg_required_else_help = true)]
-    Blob {
-        #[command(subcommand)]
-        cmd: Option<CacheBlobCmd>,
-    },
-
-    /// Read and write cached action results.
-    #[command(arg_required_else_help = true)]
-    Action {
-        #[command(subcommand)]
-        cmd: Option<CacheActionCmd>,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum CacheBlobCmd {
-    /// Store bytes from a file or stdin and print the blob digest.
-    Put {
-        /// File to store. Use `-` or omit the path to read stdin.
-        path: Option<PathBuf>,
-    },
-
-    /// Fetch blob bytes by digest.
-    Get {
-        /// Blob digest to fetch.
-        #[arg(value_parser = parse_digest)]
-        digest: Digest,
-
-        /// File to write. Defaults to stdout; use `-` for stdout.
-        #[arg(short, long)]
-        output: Option<PathBuf>,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum CacheActionCmd {
-    /// Fetch an action result by action digest.
-    Get {
-        /// Action digest to fetch.
-        #[arg(value_parser = parse_digest)]
-        action: Digest,
-    },
-
-    /// Store an action result for an action digest.
-    Put {
-        /// Action digest to store the result under.
-        #[arg(value_parser = parse_digest)]
-        action: Digest,
-
-        /// Process exit code captured for the action.
-        #[arg(long)]
-        exit_code: i32,
-
-        /// Blob digest containing captured stdout.
-        #[arg(long, value_parser = parse_digest)]
-        stdout: Digest,
-
-        /// Blob digest containing captured stderr.
-        #[arg(long, value_parser = parse_digest)]
-        stderr: Digest,
-
-        /// Declared output as `workspace/path=blob_digest`. Repeatable.
-        #[arg(long = "output", value_parser = parse_output_digest)]
-        outputs: Vec<(String, Digest)>,
-    },
-
-    /// Delete one cached action result.
-    Forget {
-        /// Action digest to remove.
-        #[arg(value_parser = parse_digest)]
-        action: Digest,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum AuthCmd {
-    /// Sign in to a provider so Fabrik can reuse its cache session.
-    Login {
-        /// Provider reference. Use `workspace` for the effective workspace provider.
-        #[arg(long)]
-        provider: String,
-
-        /// Print the authorization URL instead of opening the browser automatically.
-        #[arg(long)]
-        no_browser: bool,
-    },
-
-    /// Remove the stored session for a provider.
-    Logout {
-        /// Provider reference. Use `workspace` for the effective workspace provider.
-        #[arg(long)]
-        provider: String,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum ToolchainCmd {
-    /// Print the toolchain contract derived from mise.toml.
-    Inspect {
-        /// Mise platform key to inspect, e.g. linux-x64. Defaults to
-        /// the current host platform.
-        #[arg(long)]
-        platform: Option<String>,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum RuntimeCmd {
-    /// Serve the local runtime JSON-RPC endpoint for a session directory.
-    Rpc {
-        /// Runtime session directory containing session.json and logs.
-        session_dir: PathBuf,
-
-        /// Socket path. Defaults to `<session-dir>/control.sock`.
-        #[arg(long)]
-        socket: Option<PathBuf>,
-    },
-}
-
-#[derive(Subcommand)]
-pub enum DepsCmd {
-    /// Synchronize generated dependency artifacts from `fabrik.toml`.
-    Sync {
-        /// Sync one dependency entry by name. Defaults to all entries.
-        name: Option<String>,
-    },
-}
-
 impl Cli {
     pub fn surface_path(&self) -> Vec<&'static str> {
         self.command
@@ -488,80 +365,6 @@ impl Cmd {
     }
 }
 
-impl CacheCmd {
-    fn surface_path(&self) -> Vec<&'static str> {
-        match self {
-            Self::Stats => vec!["stats"],
-            Self::Blob { cmd } => {
-                let mut path = vec!["blob"];
-                if let Some(cmd) = cmd {
-                    path.extend(cmd.surface_path());
-                }
-                path
-            }
-            Self::Action { cmd } => {
-                let mut path = vec!["action"];
-                if let Some(cmd) = cmd {
-                    path.extend(cmd.surface_path());
-                }
-                path
-            }
-        }
-    }
-}
-
-impl CacheBlobCmd {
-    fn surface_path(&self) -> Vec<&'static str> {
-        match self {
-            Self::Put { .. } => vec!["put"],
-            Self::Get { .. } => vec!["get"],
-        }
-    }
-}
-
-impl CacheActionCmd {
-    fn surface_path(&self) -> Vec<&'static str> {
-        match self {
-            Self::Get { .. } => vec!["get"],
-            Self::Put { .. } => vec!["put"],
-            Self::Forget { .. } => vec!["forget"],
-        }
-    }
-}
-
-impl AuthCmd {
-    fn surface_path(&self) -> Vec<&'static str> {
-        match self {
-            Self::Login { .. } => vec!["login"],
-            Self::Logout { .. } => vec!["logout"],
-        }
-    }
-}
-
-impl ToolchainCmd {
-    fn surface_path(&self) -> Vec<&'static str> {
-        match self {
-            Self::Inspect { .. } => vec!["inspect"],
-        }
-    }
-}
-
-impl RuntimeCmd {
-    fn surface_path(&self) -> Vec<&'static str> {
-        match self {
-            Self::Rpc { .. } => vec!["rpc"],
-        }
-    }
-}
-
-impl DepsCmd {
-    fn surface_path(&self) -> Vec<&'static str> {
-        match self {
-            Self::Sync { .. } => vec!["sync"],
-        }
-    }
-}
-
 #[cfg(unix)]
 impl crate::commands::elixir_daemon::ElixirDaemonCmd {
     fn surface_path(&self) -> Vec<&'static str> {
@@ -581,20 +384,6 @@ fn parse_env(raw: &str) -> std::result::Result<(String, String), String> {
 
 fn parse_workspace_path(raw: &str) -> std::result::Result<WorkspacePath, String> {
     WorkspacePath::try_from(raw).map_err(|e| e.to_string())
-}
-
-fn parse_digest(raw: &str) -> std::result::Result<Digest, String> {
-    Digest::from_hex(raw).ok_or_else(|| "expected a 64-character lowercase BLAKE3 digest".into())
-}
-
-fn parse_output_digest(raw: &str) -> std::result::Result<(String, Digest), String> {
-    let (path, digest) = raw
-        .split_once('=')
-        .ok_or_else(|| "expected workspace/path=blob_digest".to_string())?;
-    if path.is_empty() {
-        return Err("output path must not be empty".into());
-    }
-    Ok((path.to_string(), parse_digest(digest)?))
 }
 
 /// Map a subprocess exit code to a CLI [`ExitCode`].
